fixed the MIDI bug (presumably)
Using a smaller buffer size of (512 instead of 2048) has made the MIDI bug impossible to
reproduce.
I put this change on the master branch and not the midi-bug-fix branch, because decreasing
the buffer size on that branch does not have the same affect. On that branch, fill_buffer is
called on flag checks in the main loop rather than interrupts for I2S tx complete. Waiting
for the loop to call those synchronously, with USB_Process() calls between them, takes way
too long with the smaller buffer size causes underflow (i.e. lots of popping)